The user can't access the certificate under some circunstances
